Refactor app, license, readme, server
