+++ src/app/app.routes.js
@@ -1,8 +1,7 @@
 const app = require("./app.middleware");
 const routes = require("../routes");
 
-app.use('/', routes.app)
+app.use("/", routes.app);
 app.use("/api/tasks", routes.tasks);
-
 
 module.exports = app;
