+++ src/server/README.md
@@ -1,0 +1,3 @@
+# App Server
+
+Here we export a funtion to initialize the server and serve the app.
